Initial commit.
An abstract data type to represent a memory-mapped Whisper file is introduced
together with functions to open and close the file, and to read the header's
metadata (from the memory-mapped data).

A sample Python script using the whisper library to display the header's
metadata of a given .wsp file is provided for comparison.