Added support to create (empty) Whisper files.
Code to write empty Whisper files is added. It writes a complete
header, including its ArchiveInfo. The rest of the file is filled
with zeroes (i.e. it is as if all the Points are (0, 0)).

The dependency on the bytestring-mmap package has been dropped: it
was necessary to write in the mmap'ed memory, which was not
feasible without modifying the package. Another package was available
on Hackage but takes directly a filename instead of a file descriptor.
(Creating a new file and mmap'ing it involves writing a spurious
byte in the file to give it its correct size before mmap'ing it.)

At this point, there are a lot of read/writeXXX functions. Some of
them should be just peek/poke (i.e. some Storable instances are
missing).