Now use a State and Writer monad for pickling.
Pickling is now done by first generating the opcodes then by serializing them.
The BINPUT opcodes are a bit more correct (the pickles generated by
pickle-dump.py and by pickle-read.hs in pickle-examples.sh are the same) but
not completely (they don't reuse the memo values).

At this time the code is not enough to serialize all the things it can
deserialize.