Mainly add a Whisper cache.
The cache mostly behaves as Data.Whisper but is in memory only. It doesn't
write itself yet on disk (using Data.Whisper) but it should.

The cache is a bit richer than Whisper itself because it permits for instance
to offer a simple Counter API, similar to EKG's.

The code isn't clean at all, in particular there is a lovely mess of String,
ByteString and Text.